refactor(worker): extract batch processor into src/worker/batch_processor
